+++ src/common/execution/request-helper.js
@@ -1,6 +1,7 @@
 const {
   checkWeb2SecretExists,
   checkWeb3SecretExists,
+  checkRequesterSecretExists,
 } = require('../sms/check');
 const { checkActiveBitInTag } = require('../utils/utils');
 const { NULL_ADDRESS, NULL_BYTES32 } = require('../utils/constant');
@@ -104,9 +105,29 @@ const checkRequestRequirements = async (
     );
     if (!isDatasetSecretSet) {
       throw Error(
-        `Dataset encryption key not set for ${dataset}. Dataset decryption will fail.`,
+        `Dataset encryption key is not set for dataset ${dataset} in the SMS. Dataset decryption will fail.`,
       );
     }
+  }
+  // check requester secrets
+  if (paramsObj[paramsKeyName.IEXEC_SECRETS]) {
+    await Promise.all(
+      Object.values(paramsObj[paramsKeyName.IEXEC_SECRETS]).map(
+        async (secretName) => {
+          const isSecetSet = await checkRequesterSecretExists(
+            contracts,
+            smsURL,
+            requester,
+            secretName,
+          );
+          if (!isSecetSet) {
+            throw Error(
+              `Requester secret "${secretName}" is not set for requester ${requester} in the SMS. Requester secret provisionning will fail.`,
+            );
+          }
+        },
+      ),
+    );
   }
   return true;
 };